requirements.txt
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.2
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.7.0
Markdown==3.2.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pytz==2020.1
requests==2.22.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
zipp==3.1.0
